requirements.txt
Django==4.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
cryptography==42.0.2
python-dotenv==1.0.0
django-cors-headers==4.3.1